--- backend/Cargo.toml
@@ -1,50 +1,0 @@
-[package]
-name = "asleepace-backend"
-version = "0.0.1"
-edition = "2021"
-publish = false
-
-[dependencies]
-cornucopia = "0.9.0"
-dotenv = "0.15.0"
-
-# For Hashing
-bcrypt = "0.9"
-rand = "0.8"
-base64 = "0.13"
-
-# Required
-postgres-types = { version = "*", features = ["derive"] }
-
-# Async
-cornucopia_async = { version = "*", features = ["with-serde_json-1"] }
-tokio = { version = "*", features = ["full"] }
-tokio-postgres = { version = "*", features = [
-    "with-serde_json-1",
-    "with-time-0_3",
-    "with-uuid-1",
-    "with-eui48-1",
-] }
-futures = "*"
-
-# Async connection pooling
-deadpool-postgres = { version = "*" }
-
-# Row serialization
-serde = { version = "*", features = ["derive"] }
-
-# Extra types
-serde_json = { version = "*", features = ["raw_value"] }
-time = "*"
-uuid = "*"
-eui48 = "*"
-rust_decimal = { version = "*", features = ["db-postgres"] }
-warp = "0.3.5"
-
-[build-dependencies]
-# Cornucopia library to automatically
-## If you're trying this example as a standalone crate, 
-## replace the path with the latest current version
-cornucopia = "0.9.0"
-serde = { version = "*", features = ["derive"] }
-serde_json = { version = "*", features = ["raw_value"] }
