--- backend/migration.sh
@@ -1,1 +1,0 @@
-cornucopia --serialize live postgres://postgres:asleepacepassword@asleepace.com:5432/master
